requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
transformers>=4.36.0
onnxruntime>=1.16.0
numpy>=1.26.0
requests>=2.31.0
redis>=5.0.0
google-genai>=1.0.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
